requirements.txt
numpy
torch>=2.7.1
tensorflow
Keras-Applications
scikit-learn
pillow>=8.0.0
transformers>=4.11.0
opencv-python>=4.12.0.88
# sortedcontainers>=2.4.0
boto3>=1.39.12
boto3-stubs>=1.39.12
botocore>=1.39.12
python-dotenv>=1.1.1